Update introduction.md
Derek via Jackie wanted the enablers listed, so each category has a list of constituent enablers.  the linnks will persist on the left for each of the categories.  Also reordered categories in the list to reflect alpha order in the left nav.  Deleted the "Digital Implementation reference in the list - deleting that category because there's nothing in it.  Fear that it will be seen as "incomplete."